+++ itolib/Patches/DoorwayPatch.cs
@@ -1,14 +1,22 @@
 using DunGen;
+using DunGen.Graph;
 using HarmonyLib;
 using itolib.Behaviours.Helpers;
+using itolib.Enums;
 using LethalLevelLoader;
 using System.Reflection;
 
 namespace itolib.Patches
 {
+    /// <summary>
+    ///     Patch for <c>SpecificDoorway</c>'s functionality.
+    /// </summary>
     [HarmonyPatch]
     internal sealed class DoorwayPatch
     {
+        /// <summary>
+        ///     Whether the current <c>Dungeon</c> is using <c>SpecificDoorway</c> or not. 
+        /// </summary>
         internal static bool? specificDoorwayActive;
 
         [HarmonyPrepare]
@@ -16,6 +24,7 @@ namespace itolib.Patches
         {
             if (original == null)
             {
+                // Reset SpecificDoorway state immediately before generation starts.
                 DungeonManager.GlobalDungeonEvents.onBeforeDungeonGenerate.AddListener(() => specificDoorwayActive = null);
             }
         }
@@ -24,20 +33,38 @@ namespace itolib.Patches
         [HarmonyPostfix]
         private static void IsValidDoorwayPairingPost(ref bool __result, DoorwayProxy a, DoorwayProxy b, TileProxy previousTile, ref float weight)
         {
+            if (!__result || specificDoorwayActive == false)
+            {
+                return;
+            }
+
+            // Check if the first Doorway that generates is a SpecificDoorway and save the result.
             specificDoorwayActive ??= a.DoorwayComponent is SpecificDoorway;
 
-            if (!__result || specificDoorwayActive == false)
+            if (specificDoorwayActive == false)
             {
                 return;
             }
 
             if (a.DoorwayComponent is SpecificDoorway doorwayA)
             {
+                // Check if Doorway should be allowed to connect, if any specified Tag is matched.
+                if (doorwayA.DoorwayConnectionTags?.Length > 0 && ValidateDoorwayTagConnection(doorwayA, b.DoorwayComponent))
+                {
+                    __result = false;
+                    weight = 0.0f;
+
+                    return;
+                }
+
+                // Check if exits should be allowed to act as entrances, if chosen to generate as such.
                 if (doorwayA.AllowSwap)
                 {
+                    // Look through every Doorway from the previous Tile.
                     foreach (DoorwayProxy previousProxy in previousTile.UsedDoorways)
                     {
-                        if (previousProxy != a && previousProxy.DoorwayComponent != null && previousProxy.DoorwayComponent is SpecificDoorway previousDoorway
+                        // Check if previous Doorway was an entrance used as an exit (inverse logic).
+                        if (previousProxy != a && previousProxy.DoorwayComponent is SpecificDoorway previousDoorway
                             && previousDoorway.AllowSwap && previousDoorway.DoorwayType == doorwayA.DoorwayType)
                         {
                             __result = false;
@@ -48,6 +75,7 @@ namespace itolib.Patches
                     }
                 }
 
+                // Check if Doorway should be allowed to connect, and apply its modified weight.
                 __result = ModifyWeights(doorwayA, next: false, ref weight);
 
                 if (!__result)
@@ -58,12 +86,29 @@ namespace itolib.Patches
 
             if (b.DoorwayComponent is SpecificDoorway doorwayB)
             {
+                // Check if Doorway should be allowed to be connected to, if any specified Tag is matched.
+                if (doorwayB.DoorwayConnectionTags?.Length > 0 && ValidateDoorwayTagConnection(doorwayB, a.DoorwayComponent))
+                {
+                    __result = false;
+                    weight = 0.0f;
+
+                    return;
+                }
+
                 __result = ModifyWeights(doorwayB, next: true, ref weight);
             }
         }
 
+        /// <summary>
+        ///     Modify a <c>SpecificDoorway</c>'s weight, based on its configuration.
+        /// </summary>
+        /// <param name="doorway"><c>SpecificDoorway</c> to modify weight for.</param>
+        /// <param name="next">Whether this <c>SpecificDoorway</c> is the next <c>Tile</c>'s entrance or not.</param>
+        /// <param name="weight">Weight for the <c>SpecificDoorway</c>.</param>
+        /// <returns>Whether this <c>SpecificDoorway</c> should be allowed to attempt to generate the path or not.</returns>
         private static bool ModifyWeights(SpecificDoorway doorway, bool next, ref float weight)
         {
+            // Check if Doorway should be allowed to generate a path, based on its configuration.
             if ((doorway.DoorwayType == DoorwayType.Neither) || (next && doorway.DoorwayType == DoorwayType.Exit && !doorway.AllowSwap)
                 || (!next && doorway.DoorwayType == DoorwayType.Entrance))
             {
@@ -72,16 +117,44 @@ namespace itolib.Patches
                 return false;
             }
 
-            if (doorway.WeightOverride != 0.0f)
+            // Apply weight override to the Doorway.
+            if (doorway.WeightOverride > 0.0f)
             {
-                weight = doorway.WeightOverride;
+                weight = doorway.WeightOverride * ((weight >= 1.0f) ? 100.0f : 1.0f); // Scale weight accordingly, if chosen as a straightened path.
             }
-            else
+
+            // Apply weight multiplier to the Doorway.
+            if (doorway.WeightMultiplier != 1.0f)
             {
                 weight *= doorway.WeightMultiplier;
             }
 
             return true;
+        }
+
+        /// <summary>
+        ///     Check if a connection beween a `SpecificDoorway` and a `Doorway` should be allowed or rejected.
+        /// </summary>
+        /// <param name="doorwayA">`SpecificDoorway` with specified tags.</param>
+        /// <param name="doorwayB">`Doorway` with tags to check.</param>
+        /// <returns>Whether the connection should be allowed or rejected.</returns>
+        private static bool ValidateDoorwayTagConnection(SpecificDoorway doorwayA, Doorway doorwayB)
+        {
+            bool tagFound = false;
+
+            // Check Doorway for any Tag that matches.
+            for (int i = 0; i < doorwayA.DoorwayConnectionTags?.Length; i++)
+            {
+                if (doorwayB.Tags.HasTag(doorwayA.DoorwayConnectionTags[i]))
+                {
+                    tagFound = true;
+
+                    break;
+                }
+            }
+
+            return doorwayA.DoorwayTagConnectionMode is DungeonFlow.TagConnectionMode.Accept ? !tagFound
+                : doorwayA.DoorwayTagConnectionMode is DungeonFlow.TagConnectionMode.Reject && tagFound;
         }
     }
 }