+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [v0.9.1]
+
+WIP
+
+- Added tag-based connection rules to `SpecificDoorway`, similar to the ones that can be set for a `Tile`.
+  - Can define one or several (`Doorway`) tags a `SpecificDoorway` should or should not be able to connect to.
+
 ## [v0.9.0]
 
 Added a couple niche scripts, reworked and overhauled various components, fixed several issues.